+++ docs/changelog.md
@@ -89,6 +89,40 @@
 
 ## 更新说明详情
 
+1.18.22 fix: fix bond_gb_zh_sina interface
+
+	1. 修复 bond_gb_zh_sina 接口
+	2. 修复 bond_gb_us_sina 接口
+
+1.18.21 fix: fix get_receipt interface
+
+	1. 修复 get_receipt 接口
+
+1.18.20 fix: fix stock_individual_spot_xq interface
+
+	1. 修复 stock_individual_spot_xq 接口
+
+1.18.19 fix: fix GitHub action
+
+	1. 修复配置文件
+
+1.18.18 fix: fix stock_individual_spot_xq interface
+
+	1. 修复 stock_individual_spot_xq 接口
+
+1.18.17 fix: fix stock_individual_spot_xq interface
+
+	1. 修复 stock_individual_spot_xq 接口
+
+1.18.16 fix: fix news_economic_baidu interface
+
+	1. 修复 news_economic_baidu 接口
+
+1.18.15 fix: fix fund_etf_scale_sse and fund_etf_scale_szse interface
+
+	1. 修复 fund_etf_scale_sse 接口-基金份额字段数据解析错误，统一使用"份"作为单位
+	2. 修复 fund_etf_scale_szse 接口-基金份额字段数据解析错误，统一使用"份"作为单位
+
 1.18.14 add: add fund_etf_scale_sse and fund_etf_scale_szse interface
 
 	1. 新增 fund_etf_scale_sse 接口-上海证券交易所 ETF 基金份额数据
@@ -589,7 +623,7 @@
 	1. 修复 stock_info_sz_delist 接口
 
 1.16.98 fix: fix fund_open_fund_info_em interface
-	
+
 	1. 修复 fund_open_fund_info_em 接口
 
 1.16.97 fix: fix fund_open_fund_info_em interface
@@ -597,7 +631,7 @@
 	1. 修复 fund_open_fund_info_em 接口
 
 1.16.96 fix: fix stock_individual_spot_xq interface
-	
+
 	1. 修复 stock_individual_spot_xq 接口
 	2. 修复 stock_cyq_em 接口
 
@@ -6084,6 +6118,20 @@
     3. 修改 air_quality_hist 接口的请求日期格式，从 '2022-01-01' 为 '20220101'
 
 ## 版本更新说明
+
+1.18.21 fix: fix get_receipt interface
+
+1.18.20 fix: fix stock_individual_spot_xq interface
+
+1.18.19 fix: fix GitHub action
+
+1.18.18 fix: fix stock_individual_spot_xq interface
+
+1.18.17 fix: fix stock_individual_spot_xq interface
+
+1.18.16 fix: fix news_economic_baidu interface
+
+1.18.15 fix: fix fund_etf_scale_sse and fund_etf_scale_szse interface
 
 1.18.14 add: add fund_etf_scale_sse and fund_etf_scale_szse interface